requirements.txt
google-generativeai
arxiv
pubmed_parser
PyYAML
requests
pytest
python-dotenv